for debugging purposes, only remove configure.in if autoconf succeeds
git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@59 fc592f45-0881-4044-91ff-c96495a54512

+++ autogen.sh
@@ -9,8 +9,7 @@ if test ! -z "$files" ; then
   cat $files >> configure.in
 fi
 echo "AC_OUTPUT(Makeconf octinst.sh)" >> configure.in
-autoconf
-rm -f configure.in
+autoconf && rm -f configure.in
 
 ## Generate ./Makeconf.in
 rm -f Makeconf.in
